Fix task 4

+++ supervised_learning/0x05-regularization/4-dropout_forward_prop.py
@@ -18,7 +18,7 @@ def dropout_forward_prop(X, weights, L, keep_prob):
         else:
             A = np.tanh(z)
             D = np.random.randn(A.shape[0], A.shape[1])
-            D = D < keep_prob
+            D = (D < keep_prob).astype(int)
             A = (np.multiply(A, D)) / keep_prob
         cache['A' + str(i + 1)] = A
     return cache
